flatten project
