+++ src/lib/blog/model.ts
@@ -1,0 +1,14 @@
+import type { Post } from '$lib/blog/types';
+import { pathToBlogs } from './constants';
+import { parse } from './parse';
+
+export const getPost = (slug: string) => {
+	const { html, metadata } = parse(pathToBlogs + `/${slug}.md`);
+	return {
+		html,
+		metadata: {
+			...metadata,
+			slug
+		}
+	} as Post;
+};
