+++ src/lib/navbar/icon.png
