--- src/routes/about/+page.svelte
@@ -1,7 +1,0 @@
-<script>
-	import Head from '$lib/seo/components/head.svelte';
-	import Navbar from '$lib/shared/client/components/navbar/navbar.svelte';
-</script>
-
-<Head title="About" />
-<Navbar selected="about" title="About" />
