+++ src/routes/blog/[slug]/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import type { PageData } from './$types';
 	import PostPage from './post.svelte';
-	import Head from '$lib/seo/components/head.svelte';
-	import Navbar from '$lib/shared/client/components/navbar/navbar.svelte';
+	import Head from '$lib/head/head.svelte';
+	import Navbar from '$lib/navbar/navbar.svelte';
 
 	export let data: PageData;
 	export const { post } = data;
