+++ src/routes/blog/[slug]/cover-photo.svelte
