+++ src/routes/featured-blogs.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { PostMetadata } from '$lib/blog/common/types';
+	import type { PostMetadata } from '$lib/blog/types';
 	import FeaturedPost from './featured-post.svelte';
 	import TopThree from './top-three.svelte';
