+++ src/routes/notes/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-	import Head from '$lib/seo/components/head.svelte';
-	import Navbar from '$lib/shared/client/components/navbar/navbar.svelte';
+	import Head from '$lib/head/head.svelte';
+	import Navbar from '$lib/navbar/navbar.svelte';
 
 	import type { PageData } from './$types';
 	import { onMount } from 'svelte';
