+++ src/routes/socials.svelte
@@ -1,6 +1,6 @@
 <script>
-	import Github from './icons/github.svelte';
-	import Linkedin from './icons/linkedin.svelte';
+	import Github from './github.svelte';
+	import Linkedin from './linkedin.svelte';
 </script>
 
 <div class="bottom-3 right-3 flex flex-col p-2 rounded bg-primary-7 fixed space-y-3">
