+++ src/routes/top-three.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { PostMetadata } from '$lib/blog/common/types';
+	import type { PostMetadata } from '$lib/blog/types';
 
 	export let posts: PostMetadata[];
 	const topThree = posts.filter((post) => post.featured !== 'true');
